Changes: --> Refactor <--
==> Remove logs
^ ^ ^ ^ ^ ^ ^ ^ ^ ^ ^ ^ ^
