+++ core/src/main/java/com/app/co/core/data/Page.kt
@@ -1,6 +1,9 @@
 package com.app.co.core.data
 
+import android.graphics.Bitmap
+
 data class Page(
     val citation: String? = null,
-    val image: String? = null
-)
+    val image: String? = null,
+    val bitmap: Bitmap
+    )
