+++ core/src/main/java/com/app/co/core/interfaces/SafeClick.kt
@@ -1,0 +1,7 @@
+package com.app.co.core.interfaces
+
+import com.app.co.core.data.Page
+
+interface SafeClick {
+    fun item(page: Page?): Page
+}
