+++ core/src/main/java/com/app/co/core/support_ext/Supports.kt
@@ -1,9 +1,12 @@
 package com.app.co.core.support_ext
 
 import android.app.Activity
+import android.content.ContentValues
 import android.content.Context
+import android.content.Intent
 import android.graphics.Bitmap
 import android.graphics.drawable.Drawable
+import android.provider.MediaStore
 import android.text.Spannable
 import android.text.SpannableString
 import android.text.style.ForegroundColorSpan
@@ -68,4 +71,36 @@ fun Any?.launchImage(
 
         }
     }
+}
+
+fun Any.share(
+    context: Context,
+    description: String,
+    bitmap: Bitmap,
+) {
+    this.toString()
+    val share = Intent(Intent.ACTION_SEND)
+    share.type = "image/jpeg"
+
+    val values = ContentValues().apply {
+        put(MediaStore.Images.Media.TITLE, "title")
+        put(MediaStore.Images.Media.MIME_TYPE, "image/jpeg")
+    }
+
+    val uri = context.contentResolver.insert(
+        MediaStore.Images.Media.EXTERNAL_CONTENT_URI,
+        values
+    )!!
+
+    try {
+        val stream = context.contentResolver.openOutputStream(uri)!!
+        bitmap.compress(Bitmap.CompressFormat.JPEG, 100, stream)
+        stream.close()
+    } catch (e: Exception) {
+        System.err.println(e.toString())
+    }
+
+    share.putExtra(Intent.EXTRA_TEXT, description)
+    share.putExtra(Intent.EXTRA_STREAM, uri)
+    context.startActivity(Intent.createChooser(share, "Compartilhar evento"))
 }
